Fix NSUInteger format specifier compiler warning
Caused by not using %lu as the format specifier for an NSUInteger, and
by not casting it to an unsigned long.

+++ xibLocalizationPostprocessor.m
@@ -50,7 +50,7 @@ int main(int argc, const char *argv[]) {
                 }
             }
             
-            NSLog(@"Warning: skipped garbage input line %d, contents: \"%@\"", lineCount, line);
+            NSLog(@"Warning: skipped garbage input line %lu, contents: \"%@\"", (unsigned long)lineCount, line);
         }
         
         if (outputStrings.length && ![outputStrings writeToFile:[NSString stringWithUTF8String:argv[2]] atomically:NO encoding:NSUTF8StringEncoding error:&error]) {
